+++ src/controllers/util/index.ts
@@ -1,5 +1,3 @@
-import fs = require('fs');
-import path = require('path');
 import { Context } from 'koa';
 import isEmpty from 'lodash/isEmpty';
 import get from 'get-value';
@@ -9,16 +7,12 @@ import {
   ProcessorTransformationRequest,
   RouterTransformationRequestData,
   RudderMessage,
-  SourceInputConversionResult,
 } from '../../types';
 import { getValueFromMessage } from '../../v0/util';
 import genericFieldMap from '../../v0/util/data/GenericFieldMapping.json';
 import { EventType, MappedToDestinationKey } from '../../constants';
-import { versionConversionFactory } from './versionConversion';
 
 export class ControllerUtility {
-  private static sourceVersionMap: Map<string, string> = new Map();
-
   public static timestampValsMap: Record<string, string[]> = {
     [EventType.IDENTIFY]: [
       `context.${RETL_TIMESTAMP}`,
@@ -28,48 +22,6 @@ export class ControllerUtility {
     ],
     [EventType.TRACK]: [`properties.${RETL_TIMESTAMP}`, ...genericFieldMap.timestamp],
   };
-
-  private static getSourceDirPath(version: string): string {
-    if (version === 'v2') {
-      return path.resolve(__dirname, '../../sources');
-    }
-    return path.resolve(__dirname, `../../${version}/sources`);
-  }
-
-  private static getSourceVersionsMap(): Map<string, any> {
-    if (this.sourceVersionMap?.size > 0) {
-      return this.sourceVersionMap;
-    }
-
-    const versions = ['v0', 'v1', 'v2'];
-
-    versions.forEach((version) => {
-      const files = fs.readdirSync(this.getSourceDirPath(version), {
-        withFileTypes: true,
-      });
-
-      const sources = files.filter((file) => file.isDirectory()).map((folder) => folder.name);
-      sources.forEach((source) => {
-        this.sourceVersionMap.set(source, version);
-      });
-    });
-    return this.sourceVersionMap;
-  }
-
-  public static adaptInputToVersion(
-    sourceType: string,
-    requestVersion: string,
-    input: NonNullable<unknown>[],
-  ): { implementationVersion: string; input: SourceInputConversionResult<NonNullable<unknown>>[] } {
-    const sourceToVersionMap = this.getSourceVersionsMap();
-    const implementationVersion = sourceToVersionMap.get(sourceType);
-
-    const conversionStrategy = versionConversionFactory.getStrategy(
-      requestVersion,
-      implementationVersion,
-    );
-    return { implementationVersion, input: conversionStrategy.convert(input) };
-  }
 
   private static getCompatibleStatusCode(status: number): number {
     return getCompatibleStatusCode(status);